Fix typo in $-zf-bp-value name
There was a warning when compiling SCSS:

DEPRECATION WARNING on line 369 of scss/util/_mixins.scss: !global
assignments won't be able to declare new variables in future versions.
Consider adding `$-zf-bp-value: null` at the top level

The reason was a typo in the assigned variable name, which had "pb"
instead of "bp".